Seperated tui from main listener
