+++ src/main/java/com/nedap/go/networking/client/GameMainClientListener.java
@@ -7,6 +7,7 @@ import com.nedap.go.model.Stone;
 import com.nedap.go.model.utils.InvalidMoveException;
 import com.nedap.go.networking.server.OnlinePlayer;
 import com.nedap.go.networking.server.utils.PlayerNotFoundException;
+import com.nedap.go.tui.GameClientTui;
 import com.nedap.go.tui.QuitGameException;
 import java.io.IOException;
 import java.io.InputStreamReader;
@@ -19,34 +20,31 @@ import java.util.Scanner;
 /**
  * The class of the TUI used for the chat application.
  */
-public class GameClientTui implements MainClientListener {
+public class GameMainClientListener implements MainClientListener {
 
   private final Scanner sc;
   private final PrintWriter output;
   private GameClient client;
-//  private String serverName = "145.126.84.96";
-  private String serverName = "localhost";
-  private int portNumber = 8080;
   private boolean isLogIn;
-  private boolean isGameStarted;
   private boolean isConnected;
   private boolean hasResigned;
   private ClientGameAdapter game;
   private GoGuiListener gui;
 
+  private GameClientTui tui;
 
-  public GameClientTui(Reader input, PrintWriter output) {
+
+  public GameMainClientListener(Reader input, PrintWriter output) {
     this.output = output;
     sc = new Scanner(input);
   }
 
-  public GameClientTui() {
+  public GameMainClientListener() {
     this(new InputStreamReader(System.in), new PrintWriter(System.out));
   }
 
-  public static void main(String[] args) {
-    GameClientTui tui = new GameClientTui();
-    tui.run();
+  public void setTui(GameClientTui tui){
+    this.tui = tui;
   }
 
   /**
@@ -71,14 +69,15 @@ public class GameClientTui implements MainClientListener {
         }
       }
     }
-    exit();
+   tui.exit();
   }
+
 
   private synchronized void play()
       throws QuitGameException, GameMismatchException, InvalidMoveException {
     while (game == null || game.isGameOver()) {
       try {
-        this.wait();
+        this.wait(500);
       } catch (InterruptedException e) {
         throw new RuntimeException(e);
       }
@@ -93,22 +92,6 @@ public class GameClientTui implements MainClientListener {
     client.sendResign();
     hasResigned = true;
     println("You resigned!");
-  }
-
-  private void exit() {
-    println("Goodbye!");
-    System.exit(0);
-  }
-
-  private void menu() {
-    String menu = """
-        Welcome to JavaGO!\s
-               Main Menu\s
-            1. Play Game\s
-            2. Help\s
-            3. Quit\s
-        """;
-    println(menu);
   }
 
   private boolean matchMakingMenu() {
@@ -146,30 +129,8 @@ public class GameClientTui implements MainClientListener {
     client.setPlayerType(Integer.parseInt(playerType));
   }
 
-  private void run() {
-    menu();
 
-    switch (getIntMenuChoice()) {
-      case 1 -> {
-        initializeClient(serverName, portNumber);
-        initializeGui();
-        sendUsername();
-        runGame();
-      }
-      case 2 -> {
-        getHelp();
-        run();
-      }
-      case 3 -> exit();
-      default -> {
-        println("Not a valid choice");
-        println("");
-        run();
-      }
-    }
-  }
-
-  private void initializeGui() {
+  public void initializeGui() {
     gui = new GoGuiListener();
     client.addListener(gui);
   }
@@ -184,45 +145,11 @@ public class GameClientTui implements MainClientListener {
     return choice;
   }
 
-  private void getHelp() {
-    OnlinePlayer player1Help = new OnlinePlayer("Player 1", Stone.BLACK);
-    OnlinePlayer player2Help = new OnlinePlayer("Player 2", Stone.WHITE);
-    GoGame helpGame = new GoGame(player1Help, player2Help);
-    try {
-      helpGame.doMove(new GoMove(player1Help, 6));
-    } catch (InvalidMoveException e) {
-      throw new RuntimeException(e);
-    }
-    String helpText = """
-        How to start a game: \s
-            1. Type the ip and the port of the host when asked.\s
-            If no host is found, you can ask again.\s
-            \s
-            2. Choose your unique username.\s
-            (If username is already in use you will need to provide a new one).\s
-            \s
-            3. Choose your player type.\s
-            After that you automatically get in the queue.\s
-            \s
-        players: \s
-            Use -H for human player.\s
-            Use -N instead of name for the Naive Strategy AI.\s
-            \s
-        How to play: \s
-              Type the preferred line number as seen in the numbering grid, \s
-              e.g. 6. \s
-        """;
-    println(helpText);
-    println(helpGame);
-    println("Press Enter");
-    sc.nextLine();
-    println("");
-  }
 
   /**
    * username is assigned by the user and communicated to the server.
    */
-  private synchronized void sendUsername() {
+  public synchronized void sendUsername(String username) {
     while (!isConnected) {
       try {
         this.wait();
@@ -230,9 +157,6 @@ public class GameClientTui implements MainClientListener {
         throw new RuntimeException(e);
       }
     }
-
-    print("Provide a username: ");
-    String username = sc.nextLine();
     client.sendUsername(username);
 
     checkUsername(username);
@@ -250,38 +174,20 @@ public class GameClientTui implements MainClientListener {
       isLogIn = true;
     } else {
       println("Username: " + username + " already exists. Choose a new one.");
-      sendUsername();
+      tui.sendUsername();
     }
   }
 
   /**
    * Initialise the client - server connection.
    */
-  private void initializeClient(String serverName, int portNumber) {
+  public void initializeClient(String serverName, int portNumber) {
     try {
       client = new GameClient(InetAddress.getByName(serverName), portNumber
           , this);
     } catch (IOException e) {
       println("Could not find host " + serverName + " @ port: " + portNumber);
-      initializeClient();
-    }
-  }
-
-  /**
-   * Initialise the client - server connection.
-   */
-  private void initializeClient() {
-    print("Please provide a server ip: ");
-    serverName = sc.nextLine();
-    print("Now please provide a port: ");
-    portNumber = sc.nextInt();
-    sc.nextLine();
-    try {
-      client = new GameClient(InetAddress.getByName(serverName),
-          portNumber, this);
-    } catch (IOException e) {
-      println("Could not find host " + serverName + " @ port: " + portNumber);
-      initializeClient();
+      tui.initializeClient();
     }
   }
 
@@ -303,7 +209,7 @@ public class GameClientTui implements MainClientListener {
   @Override
   public synchronized void connectionLost() {
     println("Disconnected from the server");
-    println("Restart client to try and reconnect");
+    println("Restart the client");
     System.exit(0);
   }
